+++ tests/e2e/entities/test_entity_crud.py
@@ -1,7 +1,7 @@
 from pytest_cases import fixture, parametrize_with_cases
 
 from pycti.test_plugin.utils import compare_values
-from tests.cases.entities import EntityTestCases
+from tests.e2e.support_cases import EntityTestCases
 
 
 @fixture
