+++ tests/e2e/support_cases/external_input_connectors.py
