+++ tests/e2e/support_cases/stream_connector.py
