+++ tests/e2e/support_cases/worker_connector.py
