+++ tests/e2e/utils/test_stix_crud.py
@@ -1,7 +1,7 @@
 from pytest_cases import fixture, parametrize_with_cases
 from stix2 import Bundle
 
-from tests.cases.entities import EntityTestCases
+from tests.e2e.support_cases import EntityTestCases
 
 
 @fixture
