+++ tests/integration/connector/test_mixin.py
